requirements.txt
fastapi==0.115.0 
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==1.26.4
opencv-python-headless==4.10.0.84
ffmpeg-python==0.2.0
